api/types: move ServiceUpdateOptions to api/types/swarm
