+++ integration/service/jobs_test.go
@@ -3,7 +3,6 @@ package service
 import (
 	"testing"
 
-	"github.com/docker/docker/api/types"
 	swarmtypes "github.com/docker/docker/api/types/swarm"
 	"github.com/docker/docker/integration/internal/swarm"
 	"gotest.tools/v3/assert"
@@ -120,7 +119,7 @@ func TestUpdateReplicatedJob(t *testing.T) {
 	spec.TaskTemplate.ForceUpdate++
 
 	_, err = client.ServiceUpdate(
-		ctx, id, service.Version, spec, types.ServiceUpdateOptions{},
+		ctx, id, service.Version, spec, swarmtypes.ServiceUpdateOptions{},
 	)
 	assert.NilError(t, err)
